+++ myems-web/src/components/MyEMS/Space/SpaceEfficiency.js
@@ -492,8 +492,6 @@ const SpaceEfficiency = ({ setRedirect, setRedirectUrl, t }) => {
     },
     [
       periodType,
-      basePeriodDateRange,
-      reportingPeriodDateRange,
       language,
       setSubmitButtonDisabled,
       setSpinnerHidden,
@@ -538,8 +536,6 @@ const SpaceEfficiency = ({ setRedirect, setRedirectUrl, t }) => {
           setSelectedSpaceName([json[0]].map(o => o.label));
           // select root space ID
           setSelectedSpaceID([json[0]].map(o => o.value));
-          // load data with root space ID
-          loadData([json[0]].map(o => o.value));
         } else {
           toast.error(t(json.description));
         }
@@ -547,7 +543,7 @@ const SpaceEfficiency = ({ setRedirect, setRedirectUrl, t }) => {
       .catch(err => {
         console.log(err);
       });
-  }, [t, loadData]);
+  }, [t]);
 
   const labelClasses = 'ls text-uppercase text-600 font-weight-semi-bold mb-0';