+++ myems-web/src/components/MyEMS/Space/SpaceSaving.js
@@ -566,8 +566,6 @@ const SpaceSaving = ({ setRedirect, setRedirectUrl, t }) => {
     },
     [
       periodType,
-      basePeriodDateRange,
-      reportingPeriodDateRange,
       language,
       setSubmitButtonDisabled,
       setSpinnerHidden,
@@ -612,8 +610,6 @@ const SpaceSaving = ({ setRedirect, setRedirectUrl, t }) => {
           setSelectedSpaceName([json[0]].map(o => o.label));
           // select root space ID
           setSelectedSpaceID([json[0]].map(o => o.value));
-          // load data with root space ID
-          loadData([json[0]].map(o => o.value));
         } else {
           toast.error(t(json.description));
         }
@@ -621,7 +617,7 @@ const SpaceSaving = ({ setRedirect, setRedirectUrl, t }) => {
       .catch(err => {
         console.log(err);
       });
-  }, [t, loadData]);
+  }, [t]);
 
   const labelClasses = 'ls text-uppercase text-600 font-weight-semi-bold mb-0';