+++ myems-web/src/components/MyEMS/Space/SpaceStatistics.js
@@ -497,8 +497,6 @@ const SpaceStatistics = ({ setRedirect, setRedirectUrl, t }) => {
     },
     [
       periodType,
-      basePeriodDateRange,
-      reportingPeriodDateRange,
       language,
       setSubmitButtonDisabled,
       setSpinnerHidden,
@@ -543,8 +541,6 @@ const SpaceStatistics = ({ setRedirect, setRedirectUrl, t }) => {
           setSelectedSpaceName([json[0]].map(o => o.label));
           // select root space ID
           setSelectedSpaceID([json[0]].map(o => o.value));
-          // load data with root space ID
-          loadData([json[0]].map(o => o.value));
         } else {
           toast.error(t(json.description));
         }
@@ -552,7 +548,7 @@ const SpaceStatistics = ({ setRedirect, setRedirectUrl, t }) => {
       .catch(err => {
         console.log(err);
       });
-  }, [t, loadData]);
+  }, [t]);
 
   const labelClasses = 'ls text-uppercase text-600 font-weight-semi-bold mb-0';