+++ CHANGELOG.md
@@ -28,6 +28,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - fixed value range issues of point in myems-admin
 - fixed pop error issues when binding working calendar to store, shopfloor, and teant in myems-admin
 - fixed space cascader reset issue of equipment energy category report in myems-web
+- fixed space cascader reset issue of space reports in myems-web
+- fixed space cascader reset issue of store energy category report in myems-web
+- fixed space cascader reset issue of tenant energy category report in myems-web
 ### Removed
 
 ## [v5.9.0] - 2025-09-28